requirements.txt
greenlet==1.1.2
nsrt-mk3-dev==1.0.0
numpy==1.22.1
pandas==1.4.0
PyAudio==0.2.11
PyMySQL==1.0.2
pyserial==3.5
python-dateutil==2.8.2
pytz==2021.3
pyusb==1.2.1
PyYAML==6.0
six==1.16.0
sounddevice-0.4.4
SQLAlchemy==1.4.31
